Refactor common code into new utility class
Code relating a mutable variable's name to its auto-generated getters and setters has been moved to a new utility class
